Move font to separate header
