+++ src/core/ui.cpp
@@ -187,27 +187,6 @@ constexpr auto quad_fragment = R"SHADER(
 
 }
 
-auto font_atlas::get_character(char c) const -> const character&
-{
-    if (auto it = chars.find(c); it != chars.end()) {
-        return it->second;
-    }
-    return missing_char;
-}
-
-auto font_atlas::length_of(std::string_view message) -> i32
-{
-    if (message.empty()) {
-        return 0;
-    }
-    i32 length = 0;
-    for (char c : message | std::views::drop(1)) {
-        length += get_character(c).advance;
-    }
-    length += get_character(message.back()).size.x;
-    return length;
-}
-
 void ui_graphics_quad::set_buffer_attributes(std::uint32_t vbo)
 {
     glBindBuffer(GL_ARRAY_BUFFER, vbo);
